Moved navigation directory in common directory.
