+++ app/src/main/java/com/koleff/kare_android/common/navigation/NavigationController.kt
@@ -1,4 +1,4 @@
-package com.koleff.kare_android.ui.navigation
+package com.koleff.kare_android.common.navigation
 
 interface NavigationController {
     suspend fun navigateTo(destination: Destination)
