+++ app/src/main/java/com/koleff/kare_android/common/navigation/NavigationNotifier.kt
@@ -1,4 +1,4 @@
-package com.koleff.kare_android.ui.navigation
+package com.koleff.kare_android.common.navigation
 
 import kotlinx.coroutines.flow.Flow
